Move selectablelist into util
