+++ libs/@hashintel/ds-components/src/components/Filter/filter-util.ts
@@ -1,4 +1,4 @@
-import { type ItemOrGroup } from "../Menu/SelectableList/selectable-list";
+import { type ItemOrGroup } from "../../util/SelectableList/selectable-list";
 
 import type { IconName } from "../Icon/icon";
