+++ libs/@hashintel/ds-components/src/components/Menu/menu.stories.tsx
@@ -1,16 +1,19 @@
 import { useCallback, useMemo, useState } from "react";
 
-import { Button } from "../Button/button";
-import { EllipsisMenu as EllipsisMenuComponent } from "./ellipsis-menu";
-import { Menu, type MenuItem } from "./menu";
-import { getItemId } from "./SelectableList/selectable-list-util";
+import { getItemId } from "../../util/SelectableList/selectable-list-util";
 import {
   defaultSelected,
   groupedItems,
   itemsWithSubActions,
-} from "./SelectableList/selectable-list.fixtures";
+} from "../../util/SelectableList/selectable-list.fixtures";
+import { Button } from "../Button/button";
+import { EllipsisMenu as EllipsisMenuComponent } from "./ellipsis-menu";
+import { Menu, type MenuItem } from "./menu";
 
-import type { Item, ItemOrGroup } from "./SelectableList/selectable-list";
+import type {
+  Item,
+  ItemOrGroup,
+} from "../../util/SelectableList/selectable-list";
 import type { Story, StoryDefault } from "@ladle/react";
 
 function prefixIds(
