+++ libs/@hashintel/ds-components/src/components/Menu/menu.tsx
@@ -3,18 +3,18 @@ import { Portal } from "@ark-ui/react/portal";
 import { cloneElement, useMemo } from "react";
 
 import { usePortalContainerRef } from "../../util/portal-container-context";
-import { type Position } from "../Tooltip/tooltip";
 import {
   SelectableList,
   type Item,
   type ItemOrGroup,
-} from "./SelectableList/selectable-list";
+} from "../../util/SelectableList/selectable-list";
 import {
   getEventHighlightedId,
   getItemId,
   isGroup,
   useLoopSelection,
-} from "./SelectableList/selectable-list-util";
+} from "../../util/SelectableList/selectable-list-util";
+import { type Position } from "../Tooltip/tooltip";
 
 export type MenuItem = Item & { selected?: boolean };
