+++ libs/@hashintel/ds-components/src/util/SelectableList/selectable-list-util.ts
@@ -1,8 +1,8 @@
 import { getTabbables, isTabbable } from "@zag-js/dom-query";
 import { useMemo } from "react";
 
-import { type Tone } from "../../../util/form-shared";
-import { type IconName } from "../../Icon/icon";
+import { type IconName } from "../../components/Icon/icon";
+import { type Tone } from "../form-shared";
 
 import type { UseMenuContext } from "@ark-ui/react/menu";
 import type { ExclusifyUnion } from "type-fest";
